Fix a bug
