+++ python/ray/_private/metrics_agent.py
@@ -31,7 +31,8 @@ from opencensus.tags import tag_value as tag_value_module
 
 import ray
 from ray._private.gcs_utils import GcsClient
-from opencensus.proto.metrics.v1.metrics_pb2 import Metric
+# from opencensus.proto.metrics.v1.metrics_pb2 import Metric
+from ray.core.generated.metrics_pb2 import Metric
 
 logger = logging.getLogger(__name__)
