+++ src/ray/protobuf/common.proto
@@ -742,6 +742,14 @@ message CoreWorkerStats {
   int64 objects_total = 24;
 }
 
+// Resource usage reported by the node reporter.
+message NodeResourceUsage {
+  // Node resource usage in serialized json.
+  // TODO: add schema for the subfields, e.g.
+  // https://github.com/ray-project/ray/blob/e54d3117a40c20ca5441d38ffc72b908bc1b0609/dashboard/modules/reporter/reporter_agent.py#L330-L347
+  string json = 1;
+}
+
 message MetricPoint {
   // Name of the metric.
   string metric_name = 1;
