+++ src/ray/protobuf/reporter.proto
@@ -77,14 +77,6 @@ message ReportOCMetricsRequest {
 
 message ReportOCMetricsReply {}
 
-// Resource usage reported by the node reporter.
-message NodeResourceUsage {
-  // Node resource usage in serialized json.
-  // TODO: add schema for the subfields, e.g.
-  // https://github.com/ray-project/ray/blob/e54d3117a40c20ca5441d38ffc72b908bc1b0609/dashboard/modules/reporter/reporter_agent.py#L330-L347
-  string json = 1;
-}
-
 // Service for communicating with the reporter agent module on a remote node.
 service ReporterService {
   // Get the profiling stats.
